deze shit

+++ GameTest/src/com/example/gametest/Beat.java
@@ -5,7 +5,7 @@ public class Beat {
 	public long endTime;
 	// intensity kan gebaseerd worden op amplitude, of later ook
 	// op de frequentiebanden waar ze in voorkomen bijvoorbeeld
-	public double intensity;
+	public float intensity;
 	
 	public long time() {
 		return (startTime + endTime) / 2;
@@ -13,13 +13,13 @@ public class Beat {
 	
 	private int length;
 	
-	public Beat(long startTime, double startIntensity) {
+	public Beat(long startTime, float startIntensity) {
 		this.startTime = endTime = startTime;
 		intensity = startIntensity;
 		length = 1;
 	}
 	
-	public void add(long time, double newIntensity) {
+	public void add(long time, float newIntensity) {
 		endTime = time;
 		length++;
 		intensity = (intensity * (length - 1) + newIntensity) / length;
